feat(migrations): add package support to inventory transfer tables, company_id to roles, and cleanup legacy workflow columns
- Added package_id and unit_id to inventory_transfer_details and inventory_transfer_shipments with appropriate foreign key constraints and indexes.
- Introduced company_id to roles table and updated unique constraints.
- Dropped unique constraint on workers table and added necessary indexes.
- Removed employee_number column from workers table.
- Added current_step column to inventory_transfers table.
- Cleaned up legacy workflow columns in inventory_transfers.
- Dropped inventory_transfer_shipments table.
- Added content column to inventory_transfers and inventory_adjustment_details tables.
- Created inventory_adjustment_details table with necessary fields and indexes.
- Updated seeders to reflect changes in product_location and removed employee_number from UserSeeder.
- Updated API routes to reflect new movement endpoints and removed legacy routes.

+++ app/Http/Resources/ProductResource.php
@@ -56,7 +56,7 @@ class ProductResource extends Resources
             'product_type' => $this->product_type,
             'is_active' => $currentLocationActive,
             'for_sale' => $this->for_sale,
-            'current_stock' => (int) $currentStock ?? 0,
+            'current_stock' => (float) $currentStock ?? 0,
             'minimum_stock' => (int) $minimumStock,
             'maximum_stock' => (int) $maximumStock,
             'created_at' => $this->created_at,
